requirements.txt
pymongo==4.15.5
robotframework==6.0.2
robotframework-appiumlibrary==3.2
robotframework-browser==16.1.0
robotframework-csvlibrary==0.0.5
robotframework-databaselibrary==1.2.4
robotframework-debuglibrary==2.3.0
robotframework-excellib==2.0.1
robotframework-faker==6.0.0
robotframework-httplibrary==0.4.2
robotframework-jsonlibrary==0.5
robotframework-lsp==1.13.0
robotframework-pabot==2.14.0
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.4
robotframework-sapguilibrary==1.1
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.7.1
robotframework-seleniumtestability==2.1.0
robotframework-soaplibrary==0.6
robotframework-stacktrace==0.4.1
robotframework-tidy==4.5.0
rpaframework-core==12.0.1
rpaframework-pdf==8.0.1
rpaframework-windows==9.0.0
